add a new function
Now if the matrix of coefficient is singular or something, you can specify
a file containing the Fock matrix of the dimer instead. This file should have
a comment line as the first line, followed by the lower-triangle format
Fock matrix of the dimer. For Gaussian users, this is a method:

%chk=your_chk_name.chk

Title

"Charge" "Multiplicity"
"Coordinates"

--Link1--

%chk=your_chk_name.chk
Geom=AllCheck Guess=Read